Move CLI code to bin folder
