+++ src/lib.rs
@@ -1,6 +1,3 @@
-pub mod app;
-pub mod utils;
-
 pub fn greeting() -> &'static str {
     "Hello"
 }
